+++ panda/src/chan/animControlCollection.h
@@ -1,16 +1,15 @@
-// Filename: animControlCollection.h
-// Created by:  drose (22Feb00)
-//
-////////////////////////////////////////////////////////////////////
-//
-// PANDA 3D SOFTWARE
-// Copyright (c) Carnegie Mellon University.  All rights reserved.
-//
-// All use of this software is subject to the terms of the revised BSD
-// license.  You should have received a copy of this license along
-// with this source code in a file named "LICENSE."
-//
-////////////////////////////////////////////////////////////////////
+/**
+ * PANDA 3D SOFTWARE
+ * Copyright (c) Carnegie Mellon University.  All rights reserved.
+ *
+ * All use of this software is subject to the terms of the revised BSD
+ * license.  You should have received a copy of this license along
+ * with this source code in a file named "LICENSE."
+ *
+ * @file animControlCollection.h
+ * @author drose
+ * @date 2000-02-22
+ */
 
 #ifndef ANIMCONTROLCOLLECTION_H
 #define ANIMCONTROLCOLLECTION_H
@@ -24,16 +23,13 @@
 
 #include "pmap.h"
 
-////////////////////////////////////////////////////////////////////
-//       Class : AnimControlCollection
-// Description : This is a named collection of AnimControl pointers.
-//               An AnimControl may be added to the collection by
-//               name.  While an AnimControl is associated, its
-//               reference count is maintained; associating a new
-//               AnimControl with the same name will decrement the
-//               previous control's reference count (and possibly
-//               delete it, unbinding its animation).
-////////////////////////////////////////////////////////////////////
+/**
+ * This is a named collection of AnimControl pointers.  An AnimControl may be
+ * added to the collection by name.  While an AnimControl is associated, its
+ * reference count is maintained; associating a new AnimControl with the same
+ * name will decrement the previous control's reference count (and possibly
+ * delete it, unbinding its animation).
+ */
 class EXPCL_PANDA_CHAN AnimControlCollection {
 PUBLISHED:
   AnimControlCollection();
@@ -51,23 +47,23 @@ PUBLISHED:
 
   void clear_anims();
 
-  // The following functions are convenience functions that vector
-  // directly into the AnimControl's functionality by anim name.
+  // The following functions are convenience functions that vector directly
+  // into the AnimControl's functionality by anim name.
 
   INLINE bool play(const string &anim_name);
-  INLINE bool play(const string &anim_name, int from, int to);
+  INLINE bool play(const string &anim_name, double from, double to);
   INLINE bool loop(const string &anim_name, bool restart);
-  INLINE bool loop(const string &anim_name, bool restart, int from, int to);
+  INLINE bool loop(const string &anim_name, bool restart, double from, double to);
   INLINE bool stop(const string &anim_name);
-  INLINE bool pose(const string &anim_name, int frame);
+  INLINE bool pose(const string &anim_name, double frame);
 
   // These functions operate on all anims at once.
   void play_all();
-  void play_all(int from, int to);
+  void play_all(double from, double to);
   void loop_all(bool restart);
-  void loop_all(bool restart, int from, int to);
+  void loop_all(bool restart, double from, double to);
   bool stop_all();
-  void pose_all(int frame);
+  void pose_all(double frame);
 
   INLINE int get_frame(const string &anim_name) const;
   INLINE int get_frame() const;
